pyproject.toml
[tool.poetry]
name = "androgee"
version = "0.1.0"
description = "A Discord bot for the egee.io Discord server"
authors = ["egee-irl <brian@egee.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
python-dotenv = "^0.17.0"
black = "^21.4b2"
pre-commit = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "androgee.androgee:start"
[tool.black]
include =  '''
/(
    androgee
)/
'''

exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.github
  | _build
  | buck-out
  | build
  | dist
  | .env
  | media
  | __pycache__
)/
'''


